# swagger.json
{"openapi":"3.1.0","info":{"title":"User & Task Management API","description":"This API manages users and their associated tasks. You can perform CRUD operations on users and tasks.","version":"1.0.0"},"paths":{"/chatbot/":{"post":{"summary":"Ask Chatbot","operationId":"ask_chatbot_chatbot__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatbotInput"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chatbot_gpt/":{"post":{"summary":"Ask Chatbot Gpt","operationId":"ask_chatbot_gpt_chatbot_gpt__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatbotInput"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/":{"post":{"summary":"Create a new user","description":"Register a new user in the system. Provide essential user information including email and phone number.","operationId":"create_user_users__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBase"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users":{"get":{"summary":"Get all users","description":"Retrieve the details of all users. You can also filter by username using the 'name' query parameter.","operationId":"get_all_users_users_get","parameters":[{"name":"name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter or get users by username.","title":"Name"},"description":"Filter or get users by username."},{"name":"email_add","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter or get users details by their email.","title":"Email Add"},"description":"Filter or get users details by their email."},{"name":"phone_num","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter or get users details by their phone number.","title":"Phone Num"},"description":"Filter or get users details by their phone number."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}":{"get":{"summary":"Get a user by ID","description":"Retrieve the user details using their unique ID. Returns 404 if not found.","operationId":"get_user_users__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update user info","description":"Modify existing user information like name, email, or phone number.","operationId":"update_user_users__user_id__put","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete a user","description":"Remove a user permanently using their ID. All tasks associated will be affected.","operationId":"delete_user_users__user_id__delete","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/":{"post":{"summary":"Create a new task","description":"Assign a new task to a user. Provide a title, content, and the user ID.","operationId":"create_task_tasks__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskBase"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks":{"get":{"summary":"Get all tasks","description":"Retrieve the details of all tasks. You can filter tasks by title, content, completion status, and the username of the user who created them.","operationId":"get_all_tasks_tasks_get","parameters":[{"name":"name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter tasks by the username of the user who created them.","title":"Name"},"description":"Filter tasks by the username of the user who created them."},{"name":"task_title","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter tasks by title.","title":"Task Title"},"description":"Filter tasks by title."},{"name":"task_content","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter tasks by content.","title":"Task Content"},"description":"Filter tasks by content."},{"name":"is_completed","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"description":"Filter tasks by completion status.","title":"Is Completed"},"description":"Filter tasks by completion status."},{"name":"user_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Filter tasks by user ID.","title":"User Id"},"description":"Filter tasks by user ID."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskBase"},"title":"Response Get All Tasks Tasks Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/{task_id}":{"get":{"summary":"Get task by ID","description":"Fetch details of a specific task using its ID. Returns task information including user ID.","operationId":"get_task_tasks__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update a task","description":"Edit the details of an existing task such as title, content, or completion status.","operationId":"update_task_tasks__task_id__put","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete a task","description":"Delete a task permanently by providing its ID. Useful for cleaning up old or completed tasks.","operationId":"delete_task_tasks__task_id__delete","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChatbotInput":{"properties":{"user_input":{"type":"string","title":"User Input"}},"type":"object","required":["user_input"],"title":"ChatbotInput"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"TaskBase":{"properties":{"title":{"type":"string","title":"Title","description":"Short title of the task."},"content":{"type":"string","title":"Content","description":"Detailed description of the task."},"user_id":{"type":"integer","title":"User Id","description":"ID of the user who owns this task."},"is_completed":{"type":"boolean","title":"Is Completed","description":"Provides whether the task has been completed or not."}},"type":"object","required":["title","content","user_id","is_completed"],"title":"TaskBase"},"TaskUpdate":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"is_completed":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Completed"}},"type":"object","required":["title","content","is_completed"],"title":"TaskUpdate"},"UserBase":{"properties":{"username":{"type":"string","title":"Username","description":"A unique username for the user."},"email":{"type":"string","title":"Email","description":"The user's email address."},"first_name":{"type":"string","title":"First Name","description":"The user's first name."},"last_name":{"type":"string","title":"Last Name","description":"The user's last name."},"phone_num":{"type":"integer","title":"Phone Num","description":"The user's phone number."}},"type":"object","required":["username","email","first_name","last_name","phone_num"],"title":"UserBase"},"UserListResponse":{"properties":{"total":{"type":"integer","title":"Total"},"users":{"items":{"$ref":"#/components/schemas/UserBase"},"type":"array","title":"Users"}},"type":"object","required":["total","users"],"title":"UserListResponse"},"UserUpdate":{"properties":{"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"phone_num":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Phone Num"}},"type":"object","required":["username","email","first_name","last_name","phone_num"],"title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}